refactor:dir 페이지별 컴포넌트 app dir 하위의 @component에서 관리되도록 시도
