+++ apps/lostark-hands-next/app/@component/notice.tsx
