+++ apps/lostark-hands-next/app/user-info/[name]/@component/item-desc.tsx
