+++ apps/lostark-hands-next/app/user-info/[name]/@component/quality-chip.tsx
