+++ apps/lostark-hands-next/app/user-info/[name]/@component/tendencies.tsx
