+++ apps/lostark-hands-next/app/user-info/[name]/avatar/@component/index.tsx
@@ -2,8 +2,8 @@
 
 import type { IArmoryAvatar, TParsedArmory } from '@/service/armories/types';
 
+import ArmoryCard from '@/app/user-info/[name]/@component/armory-card';
 import LabelLayout from '@/client-component/label-layout';
-import ArmoryCard from '@/client-component/pages/user-info/armory-card';
 
 interface IAvatarProps {
 	data: TParsedArmory<IArmoryAvatar>[][];
