+++ apps/lostark-hands-next/app/user-info/[name]/avatar/page.tsx
@@ -1,7 +1,7 @@
 import { getAvatarApi } from '@/service/armories';
 import { avatarSelector } from '@/service/armories/selector';
 
-import Avatar from '@/client-component/pages/user-info/avatar';
+import Avatar from '@/app/user-info/[name]/avatar/@component';
 
 const Page = async ({ params: { name } }: { params: { name: string } }) => {
 	const avatar = await getAvatarApi(name);
