+++ apps/lostark-hands-next/app/user-info/[name]/collection/@component/medal.tsx
@@ -2,15 +2,15 @@
 
 import type { IArmoryEquipment, TParsedArmory } from '@/service/armories/types';
 
+import ArmoryCard from '@/app/user-info/[name]/@component/armory-card';
 import LabelLayout from '@/client-component/label-layout';
 import { useModalDispatch } from '@/client-component/modal/provider';
-import ArmoryCard from '@/client-component/pages/user-info/armory-card';
 
-interface ICollectionMedalProps {
+interface IMedalProps {
 	data: TParsedArmory<IArmoryEquipment>[];
 }
 
-const CollectionMedal = ({ data }: ICollectionMedalProps) => {
+const Medal = ({ data }: IMedalProps) => {
 	const { onOpenModal } = useModalDispatch();
 
 	return (
@@ -31,4 +31,4 @@ const CollectionMedal = ({ data }: ICollectionMedalProps) => {
 	);
 };
 
-export default CollectionMedal;
+export default Medal;
