--- apps/lostark-hands-next/app/user-info/[name]/expedition/page.tsx
@@ -1,14 +1,0 @@
-import { getAvatarApi } from '@/service/armories';
-import { avatarSelector } from '@/service/armories/selector';
-
-import Avatar from '@/client-component/pages/user-info/avatar';
-
-const Page = async ({ params: { name } }: { params: { name: string } }) => {
-	const avatar = await getAvatarApi(name);
-
-	const filteredAvatar = avatarSelector(avatar);
-
-	return <Avatar data={filteredAvatar} />;
-};
-
-export default Page;
