+++ apps/lostark-hands-next/app/user-info/[name]/silblings/page.tsx
@@ -1,7 +1,7 @@
 import { getSilblingsInfoApi } from '@/service/characters';
 import { sliblingListSelector } from '@/service/characters/selector';
 
-import Silblings from '@/client-component/pages/user-info/silblings';
+import Silblings from '@/app/user-info/[name]/silblings/@component';
 
 const Page = async ({ params: { name } }: { params: { name: string } }) => {
 	const silblings = await getSilblingsInfoApi(name);
