+++ apps/lostark-hands-next/client-component/item-part-box.tsx
